Make tag pages work
These pages weren't being auto-generated because the gem we're using to
do that was stuck at an old version which didn't support auto-generating
tag or category pages.

The underlying issue is that more recent versions of jekyll-paginate-v2
are fixed at Jekyll 3.0, when we're using 4.0. It seems development has
stalled on the project and PRs to fix this are going unmerged.

As an interim fix, I've forked the last good version - 1.9.4 into our own
GH: https://github.com/openownership/jekyll-paginate-v2 and bumped the
version compatibility to 4.0 as advised in various PRs to the original.

We already pay for a Siteleaf plan which allows custom Jekyll plugins
and Siteleaf renders our content before pushing to Github (so we're not
using github pages where gems are fixed), therefore adding this gem from
our GH repo doesn't need any changes there.