Rationalise page heading and templates for it
All pages now use _includes/blocks/title.html to render their page title
and page.description for the small intro paragraph underneath it. In
addition, the intro paragraph is an actual `<p>` rather than a `<h5>`.

One additional upshot is that every page has the same bottom border
under its heading (except Guide pages which are different anyway).

This took quite some juggling, as we had 3 or 4 different template
styles and some other issues with the news page which clashed with this.