Rationalise how we include basic markdown content in pages
As with page titles, we had a few different patterns across different
pages which made it confusing to work on templates and hard to
understand where each page was truly different where it should be the
same.

This makes every single standard page (i.e. everything that's not the
guide or a blog post) behave the same way by including anything in
`page.content` immediately under the page title in the same HTML.

We don't use this content on many pages, as they have more specialised
contents such as the items in a collection, so I've also tidied up those
CMS fields to remove markdown that wasn't being displayed but would have
caused confusion.