Add news, events and blogs categories for posts
A simple way to divide these up whilst keeping everything under _posts.

Note we're reusing categories here which mean different things for
different collections, but I'll ticket and tackle that separately.

As part of this I've moved the tags for blog posts to the bottom of the
page, so that they don't get confused with categories.