Add featured blogs/news/events to the homepage
Dropping the obtusely named 'description' section where we had parked a
small custom bit of up-to-date content in favour of a selection of the
three most recent featured blog posts.

I stole the cards used elsewhere for this but added a few new box-shadow
variants to match the accent colours used for the category labels.