Replace Codekit with Gulp and Yarn
Rather than needing to use Codekit for frontend work I wanted a free,
CLI based way to compile the assets that was repeatable and worked
better with git. We already had Gulp in use through Codekit, or at least
some basic scaffolding for it, so I've  kept it for now, although we
probably should switch to webpack for future compatibility.

I've switched to Yarn to manage these packages rather than NPM/Bower
becuase it's the defacto js package manager and it's easier to have one
combined tool than two.

The generated css/js is not totally identical because I've upgraded
the compilers and minifiers along the way and moved a couple of vendored
libs to node_modules, but it makes very limited visible or functional
difference in my testing.

I've removed lightgallery in this update too, because we don't seem to
actually use it, so to clarify things it was better gone.