Generate map data from a CSV
Temporarily dropping it into the page as HTML as well, we'll need to
work on this presentation.

The JS here is a bit sloppy as I'm sketching out what I need to do, but
will probably rewrite it into a pre-processing step when we pull the
CSV from Notion automatically.