Add basic tooltips to the map page
These pluck the HTML from the country sections on the page, which have
been updated to use our 'callout' styling. I've refactored the HTML
parsing so that it's in a single place in the javascript and then drives
all of the other data for the map from a canonical set of country
objects.

The tooltips themselves are still basic, but now we can modify the HTML
version and hopefully have them 'just work' in the tooltips.