Add skeleton of filters which alter the map and country cards
This is a very basic version which just wires a select box to the
country cards and the map. To take this further I need to get the
filter values into the country data so that it's not being
parsed/processed by the javascript. Things are hard-coded in the DOM at
the moment, but it demonstrates the mechanisms for updating the map
and country cards  at least.