Tie map tooltips into selecting countries
When you select a country via click, we show the tooltip and when you
click off or deselect we hide it. In addition, that country's card is
shown on the page below (and only that country).

This has some connections to the filtering, in that we need to make sure
we retain your filters when we go de-select a country. This is a
precursor to being able to link to a country in the URL.