Store disclosure regimes in separate data CSV
There will be multiple regimes/registers per country, so we need to
keep hold of them separately and list them all, which doesn't work too
well with CSV-based data.

This might be a good case to move to yaml files instead in the future,
as that would allow us to do some better structuring and potentially
have a page per country too.