Fix commitment tracking after Notion update
We've removed a column from our internal Notion DB now that we track
commitments separately, but I hadn't updated the basic top-level check
we do to see if a country is committed to match.

I've also updated the data as part of this - it's removed a few
countries which were erroneously being included (India, Botswana,
Ecuador), presumably due to stale/bad data in the old Notion column.