Remove what-input plugin
This plugin uses window.sessionStorage to store a variable, which I
can't figure out how to turn off and is currently the only reason we'd
have to have any kind of cookie consent.

It's only an enhancement for input outlines, so isn't essential and
doesn't seem to change our current use of form inputs, so I'm removing
it.