Move governance info to connect page & rename to about
The text is so short it doesn't really need its own page at the moment,
so I've moved it. The 'connect' page was already pretty broad, so we've
also renamed it.

I fixed the temporary redirect system whilst I was doing this, so that
I could redirect the old url to the new one.